SDK 6.66.13 + skip delete warning

+++ src/main.py
@@ -92,7 +92,7 @@ def import_images(api: sly.Api, task_id: int):
 
     if g.NEED_DOWNLOAD:
         sly.fs.remove_dir(dir_=g.STORAGE_DIR)
-    if g.REMOVE_SOURCE:
+    if g.REMOVE_SOURCE and not g.IS_ON_AGENT:
         api.file.remove(team_id=g.TEAM_ID, path=g.INPUT_PATH)
         source_dir_name = g.INPUT_PATH.lstrip("/").rstrip("/")
         sly.logger.info(msg=f"Source directory: '{source_dir_name}' was successfully removed.")
